extract TD test buttons
