+++ appUI/src/pages/Systems/ChooseTDTestDialog.tsx
@@ -1,194 +1,16 @@
-import { FC, useState } from 'react';
+import { FC } from 'react';
 import { DatasetIdentifier } from '../../adapters/api/types/common.ts';
-import {
-    Button,
-    Dialog,
-    DialogActions,
-    DialogContent,
-    FormControl,
-    InputLabel,
-    MenuItem,
-    Select,
-    Stack,
-    TextField,
-    Tooltip,
-} from '@mui/material';
+import { Button, Dialog, DialogActions, DialogContent, Stack } from '@mui/material';
 import { DialogProps } from '../../adapters/types/DialogProps.ts';
 import { DialogTitleWithX } from '../../components/Mui/DialogTitle.tsx';
 import { QuestionMark } from '@mui/icons-material';
-import { useVLEAnalysisDialog } from '../../actions/VLEAnalysis/useVLEAnalysisDialog.tsx';
-import { useSlopeTestDialog } from '../../actions/Slope/useSlopeTestDialog.tsx';
-import { useGammaTestDialog } from '../../actions/Gamma/useGammaTestDialog.tsx';
-import { useRKTestDialog } from '../../actions/RK/useRKTestDialog.tsx';
-import { useHeringtonTestDialog } from '../../actions/Herington/useHeringtonTestDialog.tsx';
-import { useFredenslundTestDialog } from '../../actions/Fredenslund/useFredenslundTestDialog.tsx';
-import { useVanNessTestDialog } from '../../actions/VanNess/useVanNessTestDialog.tsx';
-import { useConfig } from '../../contexts/ConfigContext.tsx';
-import { useData } from '../../contexts/DataContext.tsx';
-
-const VLEAnalysisButton: FC<DatasetIdentifier> = (props) => {
-    const { perform, result } = useVLEAnalysisDialog(props);
-    return (
-        <>
-            <Button variant="contained" onClick={perform}>
-                Visualize data
-            </Button>
-            {result}
-        </>
-    );
-};
-
-const SlopeTestButton: FC<DatasetIdentifier> = (props) => {
-    const { perform, result } = useSlopeTestDialog(props);
-    return (
-        <>
-            <Button variant="contained" onClick={perform}>
-                Slope test
-            </Button>
-            {result}
-        </>
-    );
-};
-
-const GammaTestButton: FC<DatasetIdentifier> = (props) => {
-    const { perform, result } = useGammaTestDialog(props);
-    return (
-        <>
-            <Button variant="contained" onClick={perform}>
-                Gamma test
-            </Button>
-            {result}
-        </>
-    );
-};
-
-const RKTestButton: FC<DatasetIdentifier> = (props) => {
-    const { perform, result } = useRKTestDialog(props);
-    return (
-        <>
-            <Button variant="contained" onClick={perform}>
-                Redlich-Kister test
-            </Button>
-            {result}
-        </>
-    );
-};
-
-const HeringtonTestButton: FC<DatasetIdentifier> = (props) => {
-    const { perform, result } = useHeringtonTestDialog(props);
-    return (
-        <>
-            <Button variant="contained" onClick={perform}>
-                Herington test
-            </Button>
-            {result}
-        </>
-    );
-};
-
-const FredenslundTestButton: FC<DatasetIdentifier> = (props) => {
-    const { default_legendre_order } = useConfig();
-    const [legendre_order, setLegendre_order] = useState(default_legendre_order);
-    const { perform, result } = useFredenslundTestDialog({ ...props, legendre_order });
-    const [nextStep, setNextStep] = useState(false);
-    const sendAndReset = () => {
-        setNextStep(false);
-        perform();
-    };
-
-    const form = !nextStep ? (
-        <Button variant="contained" onClick={() => setNextStep(true)}>
-            Fredenslund test
-        </Button>
-    ) : (
-        <form onSubmit={sendAndReset}>
-            <Stack direction="row" gap={1}>
-                <TextField
-                    type="number"
-                    label="Legendre polynomial order"
-                    value={legendre_order}
-                    onChange={(e) => setLegendre_order(parseInt(e.target.value))}
-                    inputProps={{ min: 3, max: 5 }}
-                    size="small"
-                    autoFocus
-                    fullWidth
-                />
-                <Button variant="contained" onClick={sendAndReset}>
-                    Run
-                </Button>
-            </Stack>
-        </form>
-    );
-
-    return (
-        <>
-            {form}
-            {result}
-        </>
-    );
-};
-
-const VanNessTestButton: FC<DatasetIdentifier> = (props) => {
-    const { fitData } = useData();
-    const systemName = `${props.compound1}-${props.compound2}`;
-    const fitDataForSystem = fitData?.find((fit) => fit.system_name === systemName)?.fits ?? [];
-
-    const [model_name, setModel_name] = useState<string>(fitDataForSystem[0]?.model_name ?? '');
-    const { perform, result } = useVanNessTestDialog({ ...props, model_name });
-    const [nextStep, setNextStep] = useState(false);
-    const sendAndReset = () => {
-        setNextStep(false);
-        perform();
-    };
-
-    if (fitDataForSystem.length === 0)
-        return (
-            <Tooltip title="Requires a model Fitting">
-                <div>
-                    <Button fullWidth variant="contained" disabled>
-                        Van Ness test
-                    </Button>
-                </div>
-            </Tooltip>
-        );
-
-    const inputLabelId = `models-${systemName}`;
-
-    const form = !nextStep ? (
-        <Button variant="contained" onClick={() => setNextStep(true)}>
-            Van Ness test
-        </Button>
-    ) : (
-        <form onSubmit={sendAndReset}>
-            <Stack direction="row" gap={1}>
-                <FormControl fullWidth className="medium-input">
-                    <InputLabel id={inputLabelId}>Model</InputLabel>
-                    <Select
-                        size="small"
-                        labelId={inputLabelId}
-                        label="Datasets"
-                        value={model_name}
-                        onChange={(e) => setModel_name(e.target.value)}
-                    >
-                        {fitDataForSystem.map(({ model_display_name: name }) => (
-                            <MenuItem key={name} value={name} children={name} />
-                        ))}
-                    </Select>
-                </FormControl>
-                <Button variant="contained" onClick={sendAndReset}>
-                    Run
-                </Button>
-            </Stack>
-        </form>
-    );
-
-    return (
-        <>
-            {form}
-            {result}
-        </>
-    );
-};
+import { VLEAnalysisButton } from './buttons/TDTest/VLEAnalysisButton.tsx';
+import { SlopeTestButton } from './buttons/TDTest/SlopeTestButton.tsx';
+import { GammaTestButton } from './buttons/TDTest/GammaTestButton.tsx';
+import { RKTestButton } from './buttons/TDTest/RKTestButton.tsx';
+import { HeringtonTestButton } from './buttons/TDTest/HeringtonTestButton.tsx';
+import { FredenslundTestButton } from './buttons/TDTest/FredenslundTestButton.tsx';
+import { VanNessTestButton } from './buttons/TDTest/VanNessTestButton.tsx';
 
 export const ChooseTDTestDialog: FC<DialogProps & DatasetIdentifier> = ({ open, handleClose, ...props }) => (
     <Dialog open={open} onClose={handleClose} fullWidth maxWidth={'xs'}>
